Formatting

+++ project.py
@@ -1,14 +1,15 @@
-from flask import Flask, render_template, request, redirect, jsonify, url_for, flash , session
+from flask import Flask, render_template, request, redirect, abort
+from flask import jsonify, url_for, flash, session, make_response
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, CategoryItem , User
-import random , string
+from database_setup import Base, Category, CategoryItem, User
+import random
+import string
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
 import httplib2
 from flask.ext.httpauth import HTTPBasicAuth
 import json
-from flask import make_response
 import requests
 
 auth = HTTPBasicAuth()
@@ -27,64 +28,81 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 db = DBSession()
 
+
 @app.route('/')
 @app.route('/categories')
 def showCategories():
     createSession()
     categories = db.query(Category).all()
-    return render_template('categories.html',cats= categories,
+    return render_template('categories.html',
+                           cats=categories,
                            user_id=session.get('user_id'),
                            STATE=session.get('state'))
 
 
 def createSession():
     if session.get('state') is None:
-        session['state'] = ''.join(random.choice(string.ascii_uppercase + string.digits)
-                           for x in xrange(32))
+        session['state'] = ''.join(
+                                random.choice(string.ascii_uppercase +
+                                              string.digits)
+                                for x in xrange(32))
 
-@app.route('/category/<int:category_id>/edit', methods = ["GET","POST"])
+
+@app.route('/category/<int:category_id>/edit', methods=["GET", "POST"])
 def editCategory(category_id):
-    category = db.query(Category).filter_by(id = category_id).one()
+    if not verifyUser():
+        return redirect(url_for('showLogin'))
+    category = db.query(Category).filter_by(id=category_id).one()
     if request.method == "POST":
         category.name = request.form['name']
         category.description = request.form['description']
         return redirect(url_for('showCategories'))
-    return render_template("editCategory.html" , category = category)
+    return render_template("editCategory.html", category=category)
 
-@app.route('/category/<int:category_id>/items' , methods=['GET'])
+
+@app.route('/category/<int:category_id>/items', methods=['GET'])
 def showItems(category_id):
+    if not verifyUser():
+        return redirect(url_for('showLogin'))
     createSession()
-    category = db.query(Category).filter_by(id = category_id).one()
-    items = db.query(CategoryItem).filter_by(category_id = category_id)
-    return render_template('Items.html' , category = category , items = items , creator = session)
+    category = db.query(Category).filter_by(id=category_id).one()
+    items = db.query(CategoryItem).filter_by(category_id=category_id)
+    return render_template('Items.html',
+                           category=category,
+                           items=items,
+                           creator=session)
 
-@app.route('/category/<int:category_id>/items.json' , methods=['GET'])
+
+@app.route('/category/<int:category_id>/items.json', methods=['GET'])
 def showItemsJSON(category_id):
-    items = db.query(CategoryItem).filter_by(category_id = category_id).all()
+    items = db.query(CategoryItem).filter_by(category_id=category_id).all()
     return jsonify(Items=[i.serialize for i in items])
 
 
-@app.route('/category/<int:category_id>/items/add', methods=['GET' , 'POST'])
+@app.route('/category/<int:category_id>/items/add', methods=['GET', 'POST'])
 def addCatItem(category_id):
-    category = db.query(Category).filter_by(id = category_id).one()
+    if not verifyUser():
+        return redirect(url_for('showLogin'))
+    category = db.query(Category).filter_by(id=category_id).one()
     if request.method == 'POST':
-        newItem = CategoryItem(name=request.form['name'], description=request.form['description'],
-        title=request.form['title'], category_id=category_id)
+        newItem = CategoryItem(name=request.form['name'],
+                               description=request.form['description'],
+                               title=request.form['title'],
+                               category_id=category_id)
         db.add(newItem)
         db.commit()
         flash('Item Created!')
-        return redirect(url_for('showItems' , category_id = category_id))
+        return redirect(url_for('showItems', category_id=category_id))
     else:
-        return render_template('addCatItem.html' , category_id = category_id)
-
+        return render_template('addCatItem.html', category_id=category_id)
 
 
 # Edit a Category item
-
-@app.route('/category/<int:category_id>/Items/<int:catitem_id>/edit', methods=['GET', 'POST'])
+@app.route('/category/<int:category_id>/Items/<int:catitem_id>/edit',
+           methods=['GET', 'POST'])
 def editCatItem(category_id, catitem_id):
-    if 'username' not in session:
-        return redirect('/login')
+    if not verifyUser():
+        return redirect(url_for('showLogin'))
     editedItem = db.query(CategoryItem).filter_by(id=catitem_id).one()
     category = db.query(Category).filter_by(id=category_id).one()
     if request.method == 'POST':
@@ -99,43 +117,48 @@ def editCatItem(category_id, catitem_id):
         flash('Category Item Successfully Edited')
         return redirect(url_for('showItems', category_id=category_id))
     else:
-        return render_template('editCatItem.html' , item = editedItem, category_id=category_id, item_id=catitem_id)
+        return render_template('editCatItem.html',
+                               item=editedItem,
+                               category_id=category_id,
+                               item_id=catitem_id)
 
 
 # Delete a menu item
-@app.route('/category/<int:category_id>/Items/<int:catitem_id>/delete', methods=['GET', 'POST'])
+@app.route('/category/<int:category_id>/Items/<int:catitem_id>/delete',
+           methods=['GET', 'POST'])
 def deleteCatItem(category_id, catitem_id):
-    if 'username' not in session:
-        return redirect('/login')
-    restaurant = db.query(Category).filter_by(id=category_id).one()
-    itemToDelete = db.query(CategoryItem).filter_by(id=catitem_id).one()
+    if not verifyUser():
+        return redirect(url_for('showLogin'))
+    itemtodelete = db.query(CategoryItem).filter_by(id=catitem_id).one()
     if request.method == 'POST':
-        db.delete(itemToDelete)
+        db.delete(itemtodelete)
         db.commit()
         flash('Category Item Successfully Deleted')
         return redirect(url_for('showItems', category_id=category_id))
     else:
-        return render_template('deleteCatItem.html', item=itemToDelete)
+        return render_template('deleteCatItem.html', item=itemtodelete)
 
 
-@app.route('/category/<int:category_id>/delete', methods = ["GET" , "POST"])
+@app.route('/category/<int:category_id>/delete', methods=["GET", "POST"])
 def deleteCategory(category_id):
-    category = db.query(Category).filter_by(id = category_id).one()
+    if not verifyUser():
+        return redirect(url_for('showLogin'))
+    category = db.query(Category).filter_by(id=category_id).one()
     if request.method == "POST":
         db.delete(category)
         db.commit()
         flash('category %s has been deleted!' % category.name)
         return redirect(url_for('showCategories'))
-    return render_template('deleteCategory.html' , category = category)
+    return render_template('deleteCategory.html', category=category)
 
 
-@app.route('/category/add' , methods=["POST" , "GET"])
+@app.route('/category/add', methods=["POST", "GET"])
 def addCategory():
-    if session.get('user_id') is None:
-        flash('You are not unauthorized to do this action , please login first to be able to do that!')
+    if not verifyUser():
         return redirect(url_for('showLogin'))
     if request.method == "POST":
-        category = Category(name = request.form['name'] , description = request.form['description'])
+        category = Category(
+            name=request.form['name'], description=request.form['description'])
         db.add(category)
         db.commit()
         return redirect(url_for('showCategories'))
@@ -148,6 +171,8 @@ def getCategoriesJSON():
     return jsonify(categoriesItems=[i.serialize for i in categoriesItems])
 
 # Create anti-forgery state token
+
+
 @app.route('/login')
 def showLogin():
     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
@@ -161,7 +186,7 @@ def createUser():
     newuser = User(username=session['username'], email=session[
                    'email'], picture=session['picture'])
     user = db.query(User).filter_by(email=session['email']).first()
-    if not user:
+    if user is None:
         db.add(newuser)
         db.commit()
     return user.id
@@ -219,7 +244,9 @@ def gconnect():
     stored_access_token = session.get('access_token')
     stored_gplus_id = session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
+
+        response = make_response(json.dumps
+                                 ('Current user is already connected.'),
                                  200)
         response.headers['Content-Type'] = 'application/json'
         return response
@@ -247,10 +274,12 @@ def gconnect():
     output += '!</h1>'
     output += '<img src="'
     output += session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;\
+                 -webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as %s" % session['username'])
     print "done!"
     return output
+
 
 @app.route('/gdisconnect')
 def gdisconnect():
@@ -262,7 +291,8 @@ def logout():
     if session.get('access_token')is None:
         msg = 'Current user not connected'
         return msg
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % session.get('access_token')
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % session.get(
+        'access_token')
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     if result['status'] == '200':
@@ -275,63 +305,56 @@ def logout():
         del session['state']
         return redirect(url_for('showCategories'))
     else:
-        msg= 'Failed to revoke token for given user.' + session.get('access_token')
+        msg = 'Failed to revoke token for given user.' + \
+            session.get('access_token')
         return msg
 
 
 @auth.verify_password
 def verify_password(username_or_token, password):
-    #Try to see if it's a token first
+    # Try to see if it's a token first
     user_id = User.verify_auth_token(username_or_token)
     if user_id:
-        user = db.query(User).filter_by(id = user_id).one()
+        user = db.query(User).filter_by(id=user_id).one()
     else:
-        user = db.query(User).filter_by(username = username_or_token).first()
+        user = db.query(User).filter_by(username=username_or_token).first()
         if not user or not user.verify_password(password):
             return False
     g.user = user
     return True
 
 
-@app.route('/token')
-@auth.login_required
-def get_auth_token():
-    token = g.user.generate_auth_token()
-    return jsonify({'token': token.decode('ascii')})
-
-
-@app.route('/users', methods = ['POST'])
+@app.route('/users', methods=['POST'])
 def new_user():
     username = request.json.get('username')
     password = request.json.get('password')
     if username is None or password is None:
         print "missing arguments"
-        abort(400) 
-        
-    if db.query(User).filter_by(username = username).first() is not None:
+        abort(400)
+
+    if db.query(User).filter_by(username=username).first() is not None:
         print "existing user"
         user = db.query(User).filter_by(username=username).first()
-        return jsonify({'message':'user already exists'}), 200#, {'Location': url_for('get_user', id = user.id, _external = True)}
-        
-    user = User(username = username)
+        # , {'Location': url_for('get_user', id = user.id, _external = True)}
+        return jsonify({'message': 'user already exists'}), 200
+
+    user = User(username=username)
     user.hash_password(password)
     db.add(user)
     db.commit()
-    return jsonify({ 'username': user.username }), 201#, {'Location': url_for('get_user', id = user.id, _external = True)}
+    # , {'Location': url_for('get_user', id = user.id, _external = True)}
+    return jsonify({'username': user.username}), 201
 
-@app.route('/api/users/<int:id>')
-def get_user(id):
-    user = db.query(User).filter_by(id=id).one()
-    if not user:
-        abort(400)
-    return jsonify({'username': user.username})
 
-@app.route('/api/resource')
-@auth.login_required
-def get_resource():
-    return jsonify({ 'data': 'Hello, %s!' % g.user.username })
+def verifyUser():
+    if session.get('user_id') is None:
+        flash('You are not unauthorized to do this action , \
+              please login first to be able to do that!')
+        return False
+    return True
+
 
 if __name__ == "__main__":
     app.secret_key = 'super_secret_key'
     app.debug = True
-    app.run('0.0.0.0' , 5000)
+    app.run('0.0.0.0', 5000)
